Move session container to session directory
